[Validator] Moved DoctrineCache to Doctrine Bridge
