+++ src/Symfony/Component/Validator/Mapping/Cache/ApcCache.php
@@ -15,7 +15,8 @@ use Symfony\Component\Validator\Mapping\ClassMetadata;
 
 /**
  * @deprecated Deprecated since version 2.5, to be removed in 3.0.
- *             Use DoctrineCache with Doctrine\Common\Cache\ApcCache instead.
+ *             Use Symfony\Bridge\Doctrine\Validator\Mapping\DoctrineCache
+ *             with Doctrine\Common\Cache\ApcCache instead.
  */
 class ApcCache implements CacheInterface
 {
